Add documentation and missing copyright info for new files
